运维中心代码完善

+++ src/main/java/com/hoolink/sdk/enums/omc/ObsEnum.java
@@ -8,39 +8,39 @@ public enum ObsEnum {
     /**
      *创建工单
      */
-    CREATEWORK(1, "创建工单"),
+    CREATEWORK((byte)1, "创建工单"),
     /**
      *
      */
-    TRANSFER(2, "申请转派"),
+    TRANSFER((byte)2, "申请转派"),
     /**
      *
      */
-    ADDRECEIPT(3, "添加回执"),
+    ADDRECEIPT((byte)3, "添加回执"),
     /**
      *
      */
-    DELAYWORK(4, "申请延期"),
+    DELAYWORK((byte)4, "申请延期"),
     /**
      *
      */
-    CANTCOMPLETEWORK(5, "无法完成"),
+    CANTCOMPLETEWORK((byte)5, "无法完成"),
     /**
      *
      */
-    COMPLETEWORK(6, "完成工单"),
+    COMPLETEWORK((byte)6, "完成工单"),
     /**
      *
      */
-    REVIEW(7, "审批"),
+    REVIEW((byte)7, "审批"),
     ;
 
 
-    public Integer key;
+    public Byte key;
     public String value;
 
 
-    ObsEnum(Integer key, String value) {
+    ObsEnum(Byte key, String value) {
         this.key = key;
         this.value = value;
     }
